+++ local/src/main/java/com/kdjj/local/DAO/RecipeDAO.kt
@@ -1,6 +1,10 @@
-package com.kdjj.local
+package com.kdjj.local.DAO
 
 import androidx.room.*
+import com.kdjj.local.model.RecipeEntity
+import com.kdjj.local.model.RecipeMetaEntity
+import com.kdjj.local.model.RecipeStepEntity
+import com.kdjj.local.model.RecipeTypeEntity
 
 @Dao
 interface RecipeDAO {
