+++ local/src/main/java/com/kdjj/local/model/RecipeEntity.kt
@@ -1,4 +1,4 @@
-package com.kdjj.local
+package com.kdjj.local.model
 
 import androidx.room.Embedded
 import androidx.room.Relation
